+++ packages/fury-adapter-oas3-parser/test/unit/parser/oas/parseSchemaObject-test.js
@@ -533,4 +533,32 @@ describe('Schema Object', () => {
       expect(element.attributes.get('typeAttributes').toValue()).to.deep.equal(['fixedType', 'nullable']);
     });
   });
+
+  describe('#description', () => {
+    it('warns when description is not a string', () => {
+      const schema = new namespace.elements.Object({
+        description: true,
+      });
+      const parseResult = parse(context, schema);
+
+      expect(parseResult.length).to.equal(2);
+
+      expect(parseResult).to.contain.warning(
+        "'Schema Object' 'description' is not a string"
+      );
+    });
+
+    it('adds a description to the returned element', () => {
+      const schema = new namespace.elements.Object({
+        description: 'an example schema',
+      });
+      const parseResult = parse(context, schema);
+
+      expect(parseResult.length).to.equal(1);
+      expect(parseResult.get(0)).to.be.instanceof(namespace.elements.DataStructure);
+
+      const element = parseResult.get(0).content;
+      expect(element.description.toValue()).to.equal('an example schema');
+    });
+  });
 });